+++ chat/avatar/src/Avatar/ChatAvatar.tsx
