+++ chat/avatar/src/Avatar/ChatAvatar.types.ts
