+++ chat/avatar/src/ChatAvatar.spec.tsx
